+++ tests/TLDConverterTest.php
@@ -1,0 +1,87 @@
+<?php
+
+/**
+ * PHP Domain Parser: Public Suffix List based URL parsing.
+ *
+ * @see http://github.com/jeremykendall/php-domain-parser for the canonical source repository
+ *
+ * @copyright Copyright (c) 2017 Jeremy Kendall (http://jeremykendall.net)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Pdp\Tests;
+
+use Pdp\Exception;
+use Pdp\TLDConverter;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @coversDefaultClass Pdp\TopLevelDomains
+ */
+class TLDConverterTest extends TestCase
+{
+    public function testConverter()
+    {
+        $string = file_get_contents(__DIR__.'/data/root_zones.dat');
+        $res = (new TLDConverter())->convert($string);
+        self::assertInternalType('array', $res);
+        self::assertArrayHasKey('version', $res);
+        self::assertArrayHasKey('update', $res);
+        self::assertArrayHasKey('records', $res);
+    }
+
+    /**
+     * @dataProvider invalidContentProvider
+     *
+     */
+    public function testConverterThrowsException(string $content)
+    {
+        self::expectException(Exception::class);
+        (new TLDConverter())->convert($content);
+    }
+
+    /**
+     * @covers ::convert
+     * @covers ::getHeaderInfo
+     */
+    public function invalidContentProvider()
+    {
+        $double_header = <<<EOF
+# Version 2018082200, Last Updated Wed Aug 22 07:07:01 2018 UTC
+FOO
+BAR
+# Version 2018082200, Last Updated Wed Aug 22 07:07:01 2018 UTC
+ABARTH
+ABB
+ABBOTT
+ABBVIE
+EOF;
+
+        $invalid_header = <<<EOF
+# Version 2018082200
+FOO
+BAR
+EOF;
+
+        $header_no_first_line = <<<EOF
+FOO
+BAR
+# Version 2018082200, Last Updated Wed Aug 22 07:07:01 2018 UTC
+ABARTH
+ABB
+ABBOTT
+ABBVIE
+EOF;
+
+        return [
+            'double header' => [$double_header],
+            'invalid header' => [$invalid_header],
+            'empty content' => [''],
+            'header not on the first line' => [$header_no_first_line],
+        ];
+    }
+}